Prevent closed menu from being selectable
When a menu is closed, the opacity is set to 0. The height is set to 0 as well. However, due to padding top and bottom being 8, there as a 16px high invisible section present on the screen, which was selectable (the first menu item would get clicked)

To fix this, we need to remove the padding if it is in closed mode.