adding support for mouse events to IconMenu
I believe it was intended for this to work with mouse events as well as
touch events. In my project the mouse events were not working correctly
until I added this bit of code. I tested it when nested in a AppBar’s
iconElementRight and when on it’s own.