Add menu items to select fields documentation
I looked at the examples online and had a tough time understanding what
the format of `menuItems` was. I dug into the source of the page that
rendered that example and figured out what the value should be. Then I
added this code where it would be seen. I'm not sure if this is the
right format but thought it might be a step in the right direction.