[Timepicker] Redesign 24hr format as per material spec
Change the hours layour for '24hr format'. The current '24hr' format is
not in sync with the timepickers in the material design spec.

Though it is not clearly stated in the specregarding the '24hr format' oit only makes
sense to have the first 12 hours in the inner circle than in the outer one.

Closes callemall/material-ui#3766