[Timepicker] Redesign of 24hr format w.r.t material spec
Closes callemall/material-ui#3766

Reformat the layout of hours for the '24hr' format of the timepicker.
The current design has the hours 1-12 on the outer circle which is
not in sync with the timepickers from android phones which follows
material design.
Though the material design spec doesnot specify 24hr timepicker format,
it is only intuitive to have the first 12 hours on the inner circle and
the next 12 on the outer one.