[Core] Remove default theme handling from components
Having every component check for the existence of muiTheme on context,
and adding it if not found, results in significant added complexity to
the code. This reduces code clarity and maintainability, and in some
cases introduced state solely for the purpose of moving muiTheme
through the pipeline.

By removing this code we significantly simplify the components.
In addition several components are now stateless as a result.

A developer using MUI will need to pass a theme on context once in
their app, (for example with MuIThemeProvider).