Allow wrapped ListItems in selectable Lists
Provides an optional `traverse` prop for the `MakeSelectable` component wrapper. This prop specifies an array of component types that will be traversed to find nested/wrapped `ListItem` components.

Previously, all `ListItem` components needed to be first-level children of the `SelectableList` for the `MakeSelectable` functionality to work. Now, some or all `ListItem` components can be wrapped or nested in other, random components, providing additional functionality.