[TableBody](Fixes #1345) Respect child (TableRow) props when rendering
The initial issue being fixed here was the unselect all button not actually
unselecting all of the TableRow checkboxes. This was caused by the TableBody
not respecting the value passed in the props for 'selected' when a re-render
was caused. Alongside of this, I added checks for the other props that could
be passed on the TableRow (hoverable and striped).