Pass this.state.value into onDragStop & onDragStart
Someone I know was having issues getting the slider value back when the user stops sliding. It was being passed back whilst sliding (`onChange`) but that's now what they were after.

From what I can see, the only way of getting that value is creating a ref on the Slider component and accessing `this.refs.slider.getValue()` within the `onDragStop` method. Like the `onChange` method, the second param is now the slider value. I also added the same to `onDragStart` for consistency.