[CircularProgress] Remove the thickness property
We are taking advantage of the CSS animation on the next branch
to render the component. One of the main advantage is to make the
component works without a js runtime. That's quite useful when the
component is rendered server side.
However, that's making the implementation of a thickness more complex.
The material specification doesn't seems to use different thickness.
Hence, I think that it would be better to make the master branch closer
to the next one.