Removing weird label invocation.
Not sure if this is intentional, but this pattern seems very weird. I encounter a lot of strange behavior with labels and would guess that this is a bad invocation, which is trying to use the label as both a constructor and an instance.
If it is correct, could someone help me by explaining to me what this is doing? 
I'm confused somewhat by what this is doing...