[Tabs] Add new property: fixedTabWidth
The fixedTabWidth property sets a fixed width to each tab label and aligns them
to the left instead of using full width. If the prop is not set, behavior is
back to normal. This solves issue #4420.

The docs have also been updated with information and an example for this prop.