[ListItem] Add isKeyboardFocused prop
This allows the user to manually opt for the ListItem to use the
keyboard focused styling.  This is very useful for in rich text editors,
where you often want to give the appearance that something is keyboard
focused when the focus is kept in the rich text editor.

This prop allows for that to be achieved.