[TextField]Fix first character composition issue #3394
Tiny change. It seems to me that, when the component is conrolled,
'value' needs to be checked only in componentWillReceiveProps,
and event.target.value in handleInputChange has nothing to do with
hasValue state.

At least, this PR can fix #3394 issue which, still remaining despite #5540,
breaks 'first' character composition of CJK or other input methods
when used with, for example, redux-form.