[Grid] rename changed load order requiring Hidden specs to come before Grid
I presume the problem is Grid.spec imports Hidden due to testing Grid+Hidden integration, and alphabetically it now comes first (Layout was after Hidden previously).  This is fine, but this did reveal that the way we use mocha.opts is not in pure isolation, as running a single spec still loads all specs.