[eslint] Ignore no-unused-vars on object rest properties:
Often we used the object spread to do three things at the same time
when dealing with props:
- extract properties we actually needed;
- keep the rest as 'other' to pass it down;
- exclude properties in 'other'.

When excluding properties we were creating vars without using them,
therefore requiring the developer to add eslint-disable-line all over
the place.

Now eslint doesn’t warn when destructuring an object with a rest
property and unused properties.

We are also closer to Airbnb configuration.