Issue 7175
The height of the TextField was not taking into account the size of the value when the TextField was first rendered.

The code looks at the height of the hintText, or, if the value has changed, it looks at the height of the value.

Made changes so that the value is passed into the function that is calculating the height, both when the component is mounted, or the component is resized.

I have had to skip a unit test because jsdom does not support scrollHeight. The unit test code is within the test file, and could be "unskipped" later if jsdom can do layout, or another engine is used.