Allow Customization of `grey` in Palette
I've always preferred the `bluegrey` over `grey`, so was a little disappointed when I found that it was impossible to override the default.

This is a quick fix for my issue. The color will fall back on the default (like all the others), but allows for the use of other colors for `grey`.