Removing Plugin, Should be left up to implementer
Removing babel-plugin-transform-react-remove-prop-types, for a number of reasons.

Some other libraries, such as ngReact rely on the prop-types in the production build.
The plugin doesn't actually save bandwidth but actually increases the data transferred by the added bytes needed by the if statement. (although this can be changed by setting the mode to remove rather than wrap, which still doesn't solve the other issues).
Performance enhancements that affect the state of the application should be left to the implementer.