fix(BottomNavigation): allow `null` child
Prior to this commit, the `BottomNavigation` component would
fail to render if it contains a `null` child:

	TypeError: Cannot read property 'props' of null

This is because the component tries to 'preprocess' every
child, and expects every child to have `props` (see `preparedChildren`).
There are however, cases where the child is `null`, and an error will
be thrown in such cases.

It may be `null` in cases where a user would like a navigation item
to be 'hidden' if certain conditions are not fulfilled (rather
than being disabled).

This bug is not a problem in the `v1-beta` branch, but the fix
has not been applied to `master` (as of this commit, v0.19.x).

---

Closes #8859