[docs] Fork JssProvider to release the docs
The mount/unmount behavior changed between react@15 and react@16.
This is breaking the documentation page transition.

With react@15 the unmount was called first.
With react@16 the mount is called first.

I have used the easy way out: sharing the class name generator.
In the future we might also take the `generateClassName` into account
for our ref counting algo.