[ExpansionPanelSummary] Eliminate extra invocation of onClick, fix controlled mode
Fixes #9505

As [pointed out](https://github.com/mui-org/material-ui/issues/9505#issuecomment-352208990) by @oliviertassinari, [ExpansionPanelSummary](https://material-ui-next.com/api/expansion-panel-summary/) is currently invoking the provided onClick handler twice because of its use in both ButtonBase and IconButton.  I've removed the handler from the IconButton, leaving the invocation to its parent, ButtonBase.

Also noted in this issue, @hivemind9000 has [pointed out](https://github.com/mui-org/material-ui/issues/9505#issuecomment-352213633) that the `expanded` property was not behaving as expected for [ExpansionPanel](https://material-ui-next.com/api/expansion-panel/).  Currently, it uses a class instance variable, isControlled, to determine if the expanded state of the component should be controlled by the `expanded` prop.  However, this variable is only set when the component is mounted and any subsequent use of the expanded prop is ignored.

I have changed this component so that isControlled is a function that checks to see if `expanded` has been provided in props so that it can become a controlled component at any point.  This pattern is similar to the one used in Input.  I believe it is consistent with the [documentation](https://material-ui-next.com/api/expansion-panel/) for the `expanded` prop:

>If true, expands the panel, otherwise collapse it. Setting this prop enables control over the panel.