[BottomNavigationButton] onClick and onChange handler overwritten if onClick prop provided
We're not pulling the onClick prop out in the destructure, so when we apply other to ButtonBase, we overwrite the onClick handler.  The unit test is evergreen, because it's checking to see if the spy we're passing in is being called when click is simulated, which it is, though not through the handleChange method.

To fix this, I've pulled onClick before destructuring.  If onClick is provided, it will be invoked in the handleChange class method.

I've also added some tests to get BottomNavigationButton to 100% code coverage (part of the effort for #9527).