[docs] Fix 414 HTTP issue
We have done quite some iteration on this issue.
We should be in a good shape now.
I think that it would be good in the future to add a link
toward a different online editor.

@lukePeavey I haven't ported your image replacement logic.
I couldn't find it. Feel free to submit a pull-request
to add it in the future :).

Closes #9474