[Hidden] down properties are now inclusive xsDown, smDown, mdDown, lgDown, and xlDown match the behavior of their *Up* counterparts and are now inclusive of the specified breakpoint.  This is consistent between the CSS and JS implementation of this component.
Additionally, `theme.breakpoints.down` has been altered to match this behavior.  This function is now consistent with its `up` and `only` counterparts.

Documentation has been updated, as well as internal dependence on the `down` function.  The implementation of isWidthDown has been altered so that its default behavior is to be inclusive (this was hinted in an associated comment, but was not the case).

The decision to move forward with this PR was made following a conversation with @oliviertassinari and @rosskevin, which resulted in a conclusion that it would be better to deal with this inconsistency and the breaking changes its solution would require while we are in beta.

These are the following breaking changes:

- 'createBreakpoints.down' is now inclusive of the specified breakpoint
- `isWidthDown` is now inclusive of the specified breakpoint by default
- `Hidden` will include the breakpoints associated with its *Down* properties regardless of whether CSS or JS is used.