Update EnhancedButton.js
Reinstated the keyUp handler.
Added recording of the timestamp of events processed by handleClick to prevent processing of duplicated events. #9439 #9344

Background :
The handleKeyUp firing of handleClick is duplicating the event on the processing queue.  I've not been able to find the reason for the duplication, I suspect something lower down in the React stack... but I can't prove it either way.  This will however stop a duplicated click event being processed more than once.