[classes] Move the XXXClassName to the classes property
The `classes` is the standardize way to customize the className of the component.
We do no longer need to expose a long list of `XXXClassName` properties.
The next important step in this story is to have the classes key documented.
For instance, what is the `keybordFocused` key for?
Right now, people have to either look at the implementation or the behavior in their browser.
It's not ideal.

These properties were introduced before `classes`.
Using a single pattern makes things more predictable and easier to work with.

```diff
-<Tabs buttonClassName="buttonClassName" indicatorClassName="indicatorClassName" />
+<Tabs classes={{ scrollButtons: 'buttonClassName', indicator: 'indicatorClassName' }} />
```

```diff
-<TextField labelClassName="foo" formHelperTextClassName="bar" />
+<TextField InputLabelProps={{ className: 'foo' }} FormHelperTextProps={{ className: 'bar' }} />
```