[test] Use tslint instead of typescript to test definitions
While trying to to run all the *.tsx? files might show the correct usage
of this package with typescript it does not provide a complete test
suite for the definitons i.e. we cannot write tests that expect an
error.

The longterm goal is to fully extends dtslint rules and make use of
the $Expect* helpers to display where typescript can help us avoid
errors at compile time. It will also enable linting for our definitons
files.