[lab] Depend on @babel/runtime.
@babel/runtime will also be installed by the peerDependency on @material-ui/core, but it might also
be installed by some other module or by the consuming app. That means the @babel/runtime that will
actually be resolved by @material-ui/lab is not guaranteed to be the same as the one resolved by
core, without an explicit dependency.

In particular, I was installing @babel/runtime@7.0.0-rc.1 in my app, which has different file paths
from beta.42, causing imports of @babel/runtime from @material-ui/lab to fail.

This patch ensures a compatible version of the runtime is always used.