[test] add prettier to CircleCI
A file can be formatted without knowledge about other files.
For this reason we only check the format for files which were
changed compared to master. A broken format that was merged into
master will not be picked up by CI until the file was changed
again. This can be a good thing since this will not trigger CI
errors in unrelated builds and since we are only talking about
code style this a manageable tradeoff especially because formatting
can be very expensive.