Clarify on how to use the local distribution in the CONTRIBUTING file
In this commit, we are clarifying how to install the local distribution 
of Material-UI to test changes on a real case scenario.

It turns out that, if you don't point the instalation to the /build 
folder, when importing a Material-UI component, the path 
"@material-ui/core/<ComponentName>" does not work.

It took me a while to realize how to install the local distribution, 
that's why I decided to update the documentation.