Simplify the ListItem interface when href is provided
In this commit, we are implementing the change suggested in #12941.

This is a sugar sintax when the user provides the href prop, so they 
don't need to also provide button=true and component="a".

Before this change:
<ListItem href="#" button component="a" />

After this change:
<ListItem href="#" />

I'm not totally convinced this is a valuable change because it makes the 
behavior of the ListItem less predictable, but I agree that it helps 
simplify the production code.