Support type–converting when comparing selected values in SelectInput
In this commit, we are changing the comparison strategy of the
SelectInput component to match values across different types, for
example 10 and "10".

Instead of using type–converting operators (== and !=) that would
generate linter offenses, we adopted a strategy to parse the values to
strings before comparing.

This commit addresses the issue #12047