[Grid] Responsive hidden feature
Proposal to add an additional grid size. The use is very intuitive, works without a problem with Server Side Rendering. There is no need to use the Hidden component. The idea can be extended by combining the functionality with the 'false' value and an additional grid size '0' will not be needed.