fix: Set react-jss version
Latest (as of now, 10.0.0-alpha.3) has breaking changes that cause the example not to work.

in pages/_app.tsx the import path for the JssProvider is no longer valid, and proxied through the package root.  That's not a big issue and I would have just fixed that with this PR, however there is also a breaking change with the props for the provider.  Per the change log, https://github.com/cssinjs/jss/blob/master/changelog.md#breaking-changes `generateClassName` has changed to `generateId`.  Unfortunately using the same createGenerateClassName function from materialUi/core doesn't seem to work.  As the version is still marked as alpha (and updating it to be tagged as latest may have been a mistake) I think it's best to pin the version to the previous latest until version 10 is released.