Make RootRef pass along other props to children
The documentation for `RootRef` says that "Any other properties supplied will be spread to the root element (native element)." (https://material-ui.com/api/root-ref/#props), but the implementation didn't do this. This fixes that.

In case I interpreted the documentation wrong, I'd still like to propose this change regardless because it's an inexpensive quality of life change that shouldn't break any existing code.

I was using a component that would control its children via their props -- specifically an `Accordion` component that would make sure that only one `AccordionPanel` was open at a time. Something like:

```js
<Accordion maxOpen={1}>
  <AccordionPanel>Some content</AccordionPanel>
  <AccordionPanel>Some other content</AccordionPanel>
</Accordion>
```

Wrapping the `AccordionPanel` in `RootRef` breaks this -- and this PR fixes it.