Fix tooltip alignment
Fixes #15194 and #15731.

These bugs seem to be caused by the tooltip popper adding unintended spacing to the DOM just before calculating the transform, making the calculations incorrect when the actual tooltip popper style attribute is filled. By forcing the tooltip popper to never cause spacing when it enters the DOM, the calculation is now correct.