[TrapFocus] Fix error restoring focus when activeElement is null
`document.activeElement` is null in IE 11 after unmounting the element that had focus.

You can see this behavior by putting the following into an html page:

```
<input id="testInput" autofocus>
<script>
function afterTimeout() {
	console.log("afterTimeout", document.activeElement);
	var testInput = document.getElementById("testInput");
	testInput.parentNode.removeChild(testInput);
        // null in IE 11, body element in Chrome
	console.log("after removeChild", document.activeElement);
}
function handleLoaded() {
	console.log("handleLoaded", document.activeElement);
	setTimeout(afterTimeout);
}
window.addEventListener("load", handleLoaded);
</script>
```

The use case where we saw this was clicking a button in one dialog which closed that dialog (unmounting the button that was just clicked and was therefore the focus element) and then opened a new dialog. Closing the new dialog then caused an error since `lastFocus.current` was null.