[core] Fix incorrect usage of HtmlHTMLAttributes
In #9678 the types of various props were updated.

Unfortunately, the typings were not updated to with the appropriate interface.

This `HtmlHTMLAttributes` was used over the more general `HTMLAttributes` or in one case, the more
specific `ImgHTMLAttributes`.

This PR updates the `AvatarProps['imgProps']` to `React.ImgHTMLAttributes<HTMLImageElement>` as well
as using the general `HTMLAttributes` where appropriate.