[TextareaAutosize] Fix infinite rendering loop
In Firefox, the value of `scrollHeight` is unstable: in the `<TextareAutosize>` component, the `scrollHeight` of the hidden `<textarea>` sometimes changes depending on the height of the parent element. This can cause an infinite rendering loop inside the `syncHeight` handler.

Adding a `top: 0` style to the hidden `<textarea>` fixed this problem in Firefox, and the component continued to work in Chrome, IE and Safari.