[core] Let value properties of type "unknown" be type-inferred
With current typing, this Select component:

```
<Select input={<Input />} value={10} onChange={e => log(e.currentTarget.value)} />
```
has an `onChange` property of type
```
((event: React.ChangeEvent<{
    name?: string | undefined;
    value: unknown;
}>, child: React.ReactNode) => void) | undefined
```

which means `e.currentTarget.value` cannot be passed to a function taking an
argument of type `number` without adding a run-time type check
(`typeof === "number"`) or a `@ts-ignore` comment.

However, the type of `value` can be inferred using generics components, e.g.:

```
interface SelectProps<V = unknown> {
  value?: V;
}

function Select<V>(props: SelectProps<V>): JSX.Element;
```

In the example above, the `onChange` type then turns to:
```
((event: React.ChangeEvent<{
    name?: string | undefined;
    value: number | undefined;
}>, child: React.ReactNode) => void) | undefined
```

In the cases the `Select` has some `MenuItem` with value of a different type,
as in this example (`string | number`):
```
<Select input={<Input />} value={10} onChange={e => log(e.currentTarget.value)}>
  <MenuItem value="">
    <em>None</em>
  </MenuItem>
  <MenuItem value={10}>Ten</MenuItem>
  <MenuItem value={20}>Twenty</MenuItem>
  <MenuItem value={30}>Thirty</MenuItem>
</Select>
```
the `value` type can be enforced on the call site for more precise typing:
```
<Select<string | number> input={<Input />} value={10} onChange={e => log(e.currentTarget.value)} />
```
leading to an `onChange` property of type:
```
((event: React.ChangeEvent<{
    name?: string | undefined;
    value: string | number | undefined;
}>, child: React.ReactNode) => void) | undefined
```

This diff makes all component with a value property of type "unknown" generics,
with a default type to "unknown" if it cannot be inferred.

It also may opens the path to further improvements such as:
  - `TextField`: enforcing the `value` property to type `number` if the `type`
    property is `"number"`
  - `Select`: using the `children` property to infer `value` property type from
    `MenuItem` child nodes properties

Signed-off-by: Arthur Loiret <arthur.loiret@emerton-data.com>