Fix `coef` denominator in `pxToRem` function
One thing I really appreciate about MUI is the customizability and flexibility the system provides to change global defaults and have them "permeate" through the system.

Right now, the denominator is hardcoded to `14px` which assumes that my global `fontSize` is 14, which in my case, since I customized it, is wrong. This was causing some nasty "bugs" in my font sizes that I tracked down to this line. I think it make sense to have the denominator be `htmlFontSize`. If for whatever reason this does not make sense due to context I am missing, please let me know. However, this should definitely _not_ be hardcoded.