Fix link
While i guess the link  intentionally pointed to the paragraph above, it was misleading depending on the current scroll position.

When you enter the page via the search, it is already scrolled to #material-ui-pickers. The Link i edited is then in the center of the page (and therefore grabs visual attention). When clicked, it feels defunct, because nothing happens, because the page is already scrolled to the target position.

The Link text represents a git repository name, so i would expect it to link to this repository. So i think the easiest solution is to actually do so.