Update Backdrop zIndex to 1
It seems very strange to me that you would want the zIndex of the Backdrop to be -1. I would think the backdrop zIndex should be 1 so it overrides any normally positioned or fixed positioned element on the page. I am seeing strange issues with the default backdrop where some elements on the page (also MUI components) are shown above it.

The -1 seems to be the culprit.

If you aren't going to change to 1, can someone explain why -1 is desirable?

Thanks.