[useMediaQuery] hydrationCompleted true before hydrated
Fix a bug where the `hydrationCompleted` variable in `useMediaQuery()`
can be set to `true` before all components in the document are actually
hydrated.

Apply the patch suggested in #18670 that removes this optimization in
favor of more consistent rendering behavior.

Modify the test for this optimization to check for its absence, and
elsewhere, remove the use of the `testReset()` test helper function.