[core] Ensure <SelectInput> is correctly focused by 3rd party form libraries
Some 3rd party libraries like final-form-focus attempt to focus invalid inputs by calling `.focus()` on the <input> of that name.

However for <SelectInput> the <input> is a hidden input and focusing it does not result in the <SelectInput> gaining focus. Additionally the <input> does not contain any information that could be used to deduce the correct element to focus.

To deal with this we can replace the input's focus method with one that focuses the correct node. This will ensure that if a 3rd party library tries to focus the SelectInput from the dom the SelectInput actually gains focus.