[docs] Serial comma
The serial comma (also called the Oxford comma) is the last comma before coordinating conjunction and the final item in a series. Its use is considered a matter of style and is cause for much debate. Many style guides require the serial comma, and a few advise against it. There are times when it must be used to avoid confusion or ambiguity. For the sake of consistency, we recommend always using the serial comma.