[core] Add htmlFontSize to Typography interface
The Typography interface, the type returned by createTypography, doesn't
contain the property htmlFontSize. This property is always present in
the object returned by createTypography. As a result of htmlFontSize not
existing on the Typography interface, the htmlFontSize property does not
exist on the Theme type returned by createMuiTheme either.

This fix moves the htmlFontSize from FontStyleOptions back into
FontStyle where it previously existed (see: #11456)

By moving to FontStyle, the htmlFontSize property is now present on
Typography and Theme, as well as on TypographyOptions and ThemeOptions.