Include in docs directive to silence eslint erroneous warning
The `useBabelRc` hook, when used correctly as documented, makes `eslint` emit
the error:

```
React Hook "useBabelRc" cannot be called at the top level.
React Hooks must be called in a React function component or a custom React Hook function
react-hooks/rules-of-hooks
```

Users might be confused by the error and file an issue like this
https://github.com/mui-org/material-ui/issues/24008

To prevent this from happening the documentation now includes the relevant eslint
directive to disable the warning.