[TablePagination] Fix duplicated keys in rows per page
The `Select` control which appears after `Rows per page:` shows the current amount of rows per page (value, for example `10`).
When clicking on the value, a list with all of the possible options appears (option, for example `5`, `10`, `20`, etc.).
Previously, both the value and the options each had their text as the key.
This resulted in duplicated key usage, as the value `10` with key `10` has the same key as the option `10` with key `10`.
By prefixing the value with `value-` and the option with `option-`, the keys are guaranteed to be unique.