Providing a complete example on theme customization
The current theme customization documentation seems incomplete and does not give enough information to user on properly customizing the theme and other props within it. Especially the module augmentation doesn't not specify specific module names and caused a lot of frustration for me personally (and there might be others who might be feeling the same pain). I have provided above a more complete example, with specific extension points as sample for users to theme customization.