Alternative to `dir` element on `<body>` for RTL
Adding `dir="rtl"` to the `<body>` element is only one solution for configuring RTL for native components. You can also wrap them in a lower-level div, as long as it is the parent of all your UI components. This is helpful because `<body>` is often defined as a template outside the scope of a React app, whereas a `<div>` that wraps the entire application can be easily controlled from an application component like a language switching toggle.