Update module augmentation reference url
The existing URL for Module Augmentation just took you back to the section "Customization of Theme" on the TypeScript page, which only contained a link to this page, in essence a self-reference loop that added no value. I've changed the URL to the Module Augmentation section of the Declaration Merging page of the TypeScript Documentation, which is instructive as to what Module Augmentation is. I believe this link will help future MUI advanced themers understand the concept of Module Augmentation and how it relates to custom theming.