[ToggleButton] Exclusive selection value enforce
Exclusive selection for `ToggleButtonGroup` does not enforce value at the moment. Since the most common use scenario for the exclusive selection is using `ToggleButtonGroup` in a similar manner to `RadioGroup` it would be worth noting that `exclusive` flag does not enforce value in documentation for exlusive selection. Here is the issue where the change is being discussed: https://github.com/mui-org/material-ui/issues/29784