Word choice
This document has a contradictory statement in "StackOverflow is also frequented, from time to time, by the maintainers of MUI."  Something can only be "frequented" if it is visited frequently.  By contrast, "from time to time" implies the visits are only occasional, not frequent, and perhaps closer to seldom.  I'm assuming the intent was to suggest MUI maintainers only occasionally visit, so I propose "visited" in place of "frequented," but if the truth is that MUI maintainers frequently visit, then perhaps just drop the "from time to time" altogether instead.