[docs] Update TextField multiline description
The documentation here read to me that if you set `multiline` and use the `rows` prop, you'd get a textarea component that was not a `TextareaAutosize` component. Looking into the code it looks like that used to be the case but was changed in this PR: git@github.com:jontewks/material-ui.git to always be a `TextareaAutosize` no matter what props are passed. Updated the documentation wording here to be more clear that using the `multiline` prop will always result in a `TextareaAutosize` component, and then any additional props just bound how that `TextareaAutosize` adjusts.