[Link} Allow null color
Allowing `null` for `color` would provide the same functionality that a value of `initial` used to provide which is to avoid adding any color styles and therefore get the default link color for the browser. `inherit` will explicitly set the color to the current **text** color which would generally be different than the default link color (e.g. black vs. blue).