[Divider] Avoid adding aria semantics if the divider has children
Elements with role=separator can only have children of role=presentational (or all children are converted automatically to role=presentational). This means that placing a heading as a children doesn't work. Right now there is no way of removing the role afterwards, because it is always applied too dividers with children.

With this change it is not applied automatically if children are present, but can still be applied manually.

Closes #33351