Standardizing Darken Function To Output HSL
I've added an RGBToHSL function and modified decompose colors to output HSL to standardize the `darken` function, I tested `lighten` and it works fine.

I tried to follow the existing code format as best I could. I think the next step would be to adding exports and tests if this is an acceptable fix. 

Signed-off-by: Mark Daniel <83738282+Mark-777-0@users.noreply.github.com>