[FocusTrap] Prevent restoring focus while opening a dialog/modal
Even though the beginning of the restore-focus effect returns if `!open`, the cleanup function will still execute when `open` changes from `false` to `true`. In strict mode with the dev react build, it is then possible for `nodeToRestore.current` to be non-null and return focus prematurely (during open rather than during close).