Correctly format code in component `Snackbar` prop `key` JSDoc description.
Inline code examples in JSDoc descriptions should be contained within markdown backticks. This ensures they render correctly as code in editor IntelliSense as well as in the generated Material UI API docs page. This is especially important if the code example looks like a JSX or HTML tag, otherwise markdown renderers will attempt to render it as HTML. If the renderer only accepts a subset of HTML tags (such as VS Code IntelliSense) then it will strip the code example out of the rendered HTML. In other situations where React is used to render the markdown without an allowlist for JSX/HTML (e.g. the Storybook docs page props table), then not wrapping JSX code examples in backticks in the source JSDoc description markdown can cause a React render error.